Reorganize test dependencies. * so that analysis-api does not depend on compiler anymore.
(cherry picked from commit a9e90084299b4b9543df57a93dd04a13695f6fca)
