+++ compiler-plugin/build.gradle.kts
@@ -1,4 +1,3 @@
-import com.google.devtools.ksp.RelativizingPathProvider
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 evaluationDependsOn(":common-util")
@@ -7,12 +6,6 @@ description = "Kotlin Symbol Processing"
 
 val intellijVersion: String by project
 val kotlinBaseVersion: String by project
-val junitVersion: String by project
-val junit5Version: String by project
-val junitPlatformVersion: String by project
-
-val libsForTesting by configurations.creating
-val libsForTestingCommon by configurations.creating
 
 tasks.withType<KotlinCompile> {
     compilerOptions.freeCompilerArgs.add("-Xjvm-default=all-compatibility")
@@ -55,73 +48,6 @@ dependencies {
 
     implementation(project(":api"))
     implementation(project(":common-util"))
-
-    testImplementation(kotlin("stdlib", kotlinBaseVersion))
-    testImplementation("org.jetbrains.kotlin:kotlin-compiler:$kotlinBaseVersion")
-    testImplementation("org.jetbrains.kotlin:kotlin-compiler-internal-test-framework:$kotlinBaseVersion")
-    testImplementation("org.jetbrains.kotlin:kotlin-scripting-compiler:$kotlinBaseVersion")
-
-    testImplementation("junit:junit:$junitVersion")
-    testImplementation("org.junit.jupiter:junit-jupiter-api:$junit5Version")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:$junit5Version")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-params:$junit5Version")
-    testRuntimeOnly("org.junit.platform:junit-platform-suite:$junitPlatformVersion")
-
-    testImplementation(project(":test-utils"))
-
-    libsForTesting(kotlin("stdlib", kotlinBaseVersion))
-    libsForTesting(kotlin("test", kotlinBaseVersion))
-    libsForTesting(kotlin("script-runtime", kotlinBaseVersion))
-    libsForTestingCommon(kotlin("stdlib-common", kotlinBaseVersion))
-}
-
-tasks.register<Copy>("CopyLibsForTesting") {
-    from(configurations.get("libsForTesting"))
-    into("dist/kotlinc/lib")
-    val escaped = Regex.escape(kotlinBaseVersion)
-    rename("(.+)-$escaped\\.jar", "$1.jar")
-}
-
-tasks.register<Copy>("CopyLibsForTestingCommon") {
-    from(configurations.get("libsForTestingCommon"))
-    into("dist/common")
-    val escaped = Regex.escape(kotlinBaseVersion)
-    rename("(.+)-$escaped\\.jar", "$1.jar")
-}
-
-fun Project.javaPluginConvention(): JavaPluginConvention = the()
-val JavaPluginConvention.testSourceSet: SourceSet
-    get() = sourceSets.getByName("test")
-val Project.testSourceSet: SourceSet
-    get() = javaPluginConvention().testSourceSet
-
-tasks.test {
-    dependsOn("CopyLibsForTesting")
-    dependsOn("CopyLibsForTestingCommon")
-    maxHeapSize = "2g"
-
-    useJUnitPlatform()
-
-    systemProperty("idea.is.unit.test", "true")
-    systemProperty("java.awt.headless", "true")
-    environment("NO_FS_ROOTS_ACCESS_CHECK", "true")
-
-    testLogging {
-        events("passed", "skipped", "failed")
-    }
-
-    lateinit var tempTestDir: File
-    doFirst {
-        val ideaHomeDir = buildDir.resolve("tmp/ideaHome").takeIf { it.exists() || it.mkdirs() }!!
-        jvmArgumentProviders.add(RelativizingPathProvider("idea.home.path", ideaHomeDir))
-
-        tempTestDir = createTempDir()
-        jvmArgumentProviders.add(RelativizingPathProvider("java.io.tmpdir", tempTestDir))
-    }
-
-    doLast {
-        delete(tempTestDir)
-    }
 }
 
 val dokkaJavadocJar by tasks.register<Jar>("dokkaJavadocJar") {
