+++ test-utils/src/test/kotlin/com/google/devtools/ksp/test/AbstractKSPAATest.kt
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.google.devtools.ksp.impl.test
+package com.google.devtools.ksp.test
 
 import com.google.devtools.ksp.impl.CommandLineKSPLogger
 import com.google.devtools.ksp.impl.KSPJvmConfig
